+++ react/typings/global.d.ts
@@ -22,7 +22,7 @@ declare global {
     orientation?: Orientations
     specificationsOptions?: ConditionWithName[]
     specificationName?: string
-    specificationValueSeparator?: string
+    multipleValuesSeparator?: string
   }
 
   type ConditionWithName = Condition & {